phase 1 & 2
