+++ logservice/logpuller/region_event_entries.go
@@ -1,0 +1,124 @@
+// Copyright 2024 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package logpuller
+
+import (
+	"time"
+
+	"github.com/pingcap/kvproto/pkg/cdcpb"
+	"github.com/pingcap/log"
+	"github.com/pingcap/ticdc/pkg/common"
+	"github.com/pingcap/ticdc/pkg/spanz"
+	"github.com/pingcap/ticdc/pkg/util"
+	"go.uber.org/zap"
+)
+
+func handleEventEntries(span *subscribedSpan, state *regionFeedState, entries *cdcpb.Event_Entries_) {
+	regionID, _, _ := state.getRegionMeta()
+	state.updateRuntimeLastEvent(time.Now())
+	assembleRowEvent := func(regionID uint64, entry *cdcpb.Event_Row) common.RawKVEntry {
+		var opType common.OpType
+		switch entry.GetOpType() {
+		case cdcpb.Event_Row_DELETE:
+			opType = common.OpTypeDelete
+		case cdcpb.Event_Row_PUT:
+			opType = common.OpTypePut
+		default:
+			log.Panic("meet unknown op type", zap.Any("entry", entry))
+		}
+		return common.RawKVEntry{
+			OpType:   opType,
+			Key:      entry.Key,
+			Value:    entry.GetValue(),
+			StartTs:  entry.StartTs,
+			CRTs:     entry.CommitTs,
+			RegionID: regionID,
+			OldValue: entry.GetOldValue(),
+		}
+	}
+
+	for _, entry := range entries.Entries.GetEntries() {
+		switch entry.Type {
+		case cdcpb.Event_INITIALIZED:
+			state.setInitialized()
+			state.markRuntimeReplicating(time.Now())
+			log.Debug("region is initialized",
+				zap.Int64("tableID", span.span.TableID),
+				zap.Uint64("regionID", regionID),
+				zap.Uint64("requestID", state.requestID),
+				zap.String("startKey", spanz.HexKey(span.span.StartKey)),
+				zap.String("endKey", spanz.HexKey(span.span.EndKey)))
+			for _, cachedEvent := range state.matcher.matchCachedRow(true) {
+				span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, cachedEvent))
+			}
+			state.matcher.matchCachedRollbackRow(true)
+		case cdcpb.Event_COMMITTED:
+			resolvedTs := state.getLastResolvedTs()
+			if entry.CommitTs <= resolvedTs {
+				log.Fatal("The CommitTs must be greater than the resolvedTs",
+					zap.Int64("tableID", span.span.TableID),
+					zap.Uint64("regionID", regionID),
+					zap.Uint64("requestID", state.requestID),
+					zap.String("EventType", "COMMITTED"),
+					zap.Uint64("CommitTs", entry.CommitTs),
+					zap.Uint64("resolvedTs", resolvedTs),
+					zap.String("key", util.RedactKey(entry.GetKey())))
+			}
+			span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, entry))
+		case cdcpb.Event_PREWRITE:
+			state.matcher.putPrewriteRow(entry)
+		case cdcpb.Event_COMMIT:
+			// NOTE: matchRow should always be called even if the event is stale.
+			if !state.matcher.matchRow(entry, state.isInitialized()) {
+				if !state.isInitialized() {
+					state.matcher.cacheCommitRow(entry)
+					continue
+				}
+				log.Fatal("prewrite not match",
+					zap.Int64("tableID", span.span.TableID),
+					zap.Uint64("regionID", state.getRegionID()),
+					zap.Uint64("requestID", state.requestID),
+					zap.Uint64("startTs", entry.GetStartTs()),
+					zap.Uint64("commitTs", entry.GetCommitTs()),
+					zap.String("key", util.RedactKey(entry.GetKey())))
+			}
+
+			// TiKV can send events with StartTs/CommitTs less than startTs.
+			isStaleEvent := entry.CommitTs <= span.startTs
+			if isStaleEvent {
+				continue
+			}
+
+			// NOTE: state.getLastResolvedTs() will never less than startTs.
+			resolvedTs := state.getLastResolvedTs()
+			if entry.CommitTs <= resolvedTs {
+				log.Fatal("The CommitTs must be greater than the resolvedTs",
+					zap.Int64("tableID", span.span.TableID),
+					zap.Uint64("regionID", regionID),
+					zap.Uint64("requestID", state.requestID),
+					zap.String("EventType", "COMMIT"),
+					zap.Uint64("CommitTs", entry.CommitTs),
+					zap.Uint64("resolvedTs", resolvedTs),
+					zap.String("key", util.RedactKey(entry.GetKey())))
+			}
+			span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, entry))
+		case cdcpb.Event_ROLLBACK:
+			if !state.isInitialized() {
+				state.matcher.cacheRollbackRow(entry)
+				continue
+			}
+			state.matcher.rollbackRow(entry)
+		}
+	}
+}
