+++ logservice/logpuller/region_event_handler.go
@@ -19,12 +19,8 @@ import (
 
 	"github.com/pingcap/kvproto/pkg/cdcpb"
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/pkg/common"
 	"github.com/pingcap/ticdc/pkg/metrics"
-	"github.com/pingcap/ticdc/pkg/spanz"
-	"github.com/pingcap/ticdc/pkg/util"
 	"github.com/pingcap/ticdc/utils/dynstream"
-	"github.com/tikv/client-go/v2/oracle"
 	"go.uber.org/zap"
 )
 
@@ -245,175 +241,4 @@ func (h *regionEventHandler) handleRegionError(state *regionFeedState) {
 			zap.Bool("reschedule", removed),
 			zap.Error(failure.err))
 	}
-}
-
-func handleEventEntries(span *subscribedSpan, state *regionFeedState, entries *cdcpb.Event_Entries_) {
-	regionID, _, _ := state.getRegionMeta()
-	state.updateRuntimeLastEvent(time.Now())
-	assembleRowEvent := func(regionID uint64, entry *cdcpb.Event_Row) common.RawKVEntry {
-		var opType common.OpType
-		switch entry.GetOpType() {
-		case cdcpb.Event_Row_DELETE:
-			opType = common.OpTypeDelete
-		case cdcpb.Event_Row_PUT:
-			opType = common.OpTypePut
-		default:
-			log.Panic("meet unknown op type", zap.Any("entry", entry))
-		}
-		return common.RawKVEntry{
-			OpType:   opType,
-			Key:      entry.Key,
-			Value:    entry.GetValue(),
-			StartTs:  entry.StartTs,
-			CRTs:     entry.CommitTs,
-			RegionID: regionID,
-			OldValue: entry.GetOldValue(),
-		}
-	}
-
-	for _, entry := range entries.Entries.GetEntries() {
-		switch entry.Type {
-		case cdcpb.Event_INITIALIZED:
-			state.setInitialized()
-			state.markRuntimeReplicating(time.Now())
-			log.Debug("region is initialized",
-				zap.Int64("tableID", span.span.TableID),
-				zap.Uint64("regionID", regionID),
-				zap.Uint64("requestID", state.requestID),
-				zap.String("startKey", spanz.HexKey(span.span.StartKey)),
-				zap.String("endKey", spanz.HexKey(span.span.EndKey)))
-			for _, cachedEvent := range state.matcher.matchCachedRow(true) {
-				span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, cachedEvent))
-			}
-			state.matcher.matchCachedRollbackRow(true)
-		case cdcpb.Event_COMMITTED:
-			resolvedTs := state.getLastResolvedTs()
-			if entry.CommitTs <= resolvedTs {
-				log.Fatal("The CommitTs must be greater than the resolvedTs",
-					zap.Int64("tableID", span.span.TableID),
-					zap.Uint64("regionID", regionID),
-					zap.Uint64("requestID", state.requestID),
-					zap.String("EventType", "COMMITTED"),
-					zap.Uint64("CommitTs", entry.CommitTs),
-					zap.Uint64("resolvedTs", resolvedTs),
-					zap.String("key", util.RedactKey(entry.GetKey())))
-			}
-			span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, entry))
-		case cdcpb.Event_PREWRITE:
-			state.matcher.putPrewriteRow(entry)
-		case cdcpb.Event_COMMIT:
-			// NOTE: matchRow should always be called even if the event is stale.
-			if !state.matcher.matchRow(entry, state.isInitialized()) {
-				if !state.isInitialized() {
-					state.matcher.cacheCommitRow(entry)
-					continue
-				}
-				log.Fatal("prewrite not match",
-					zap.Int64("tableID", span.span.TableID),
-					zap.Uint64("regionID", state.getRegionID()),
-					zap.Uint64("requestID", state.requestID),
-					zap.Uint64("startTs", entry.GetStartTs()),
-					zap.Uint64("commitTs", entry.GetCommitTs()),
-					zap.String("key", util.RedactKey(entry.GetKey())))
-			}
-
-			// TiKV can send events with StartTs/CommitTs less than startTs.
-			isStaleEvent := entry.CommitTs <= span.startTs
-			if isStaleEvent {
-				continue
-			}
-
-			// NOTE: state.getLastResolvedTs() will never less than startTs.
-			resolvedTs := state.getLastResolvedTs()
-			if entry.CommitTs <= resolvedTs {
-				log.Fatal("The CommitTs must be greater than the resolvedTs",
-					zap.Int64("tableID", span.span.TableID),
-					zap.Uint64("regionID", regionID),
-					zap.Uint64("requestID", state.requestID),
-					zap.String("EventType", "COMMIT"),
-					zap.Uint64("CommitTs", entry.CommitTs),
-					zap.Uint64("resolvedTs", resolvedTs),
-					zap.String("key", util.RedactKey(entry.GetKey())))
-			}
-			span.kvEventsCache = append(span.kvEventsCache, assembleRowEvent(regionID, entry))
-		case cdcpb.Event_ROLLBACK:
-			if !state.isInitialized() {
-				state.matcher.cacheRollbackRow(entry)
-				continue
-			}
-			state.matcher.rollbackRow(entry)
-		}
-	}
-}
-
-func handleResolvedTs(span *subscribedSpan, state *regionFeedState, resolvedTs uint64) uint64 {
-	if state.isStale() || !state.isInitialized() {
-		return 0
-	}
-	state.matcher.tryCleanUnmatchedValue()
-	regionID := state.getRegionID()
-	lastResolvedTs := state.getLastResolvedTs()
-	if resolvedTs < lastResolvedTs {
-		log.Debug("The resolvedTs is fallen back in subscription client",
-			zap.Uint64("subscriptionID", uint64(state.region.subscribedSpan.subID)),
-			zap.Uint64("regionID", regionID),
-			zap.Uint64("resolvedTs", resolvedTs),
-			zap.Uint64("lastResolvedTs", lastResolvedTs))
-		return 0
-	}
-
-	state.updateResolvedTs(resolvedTs)
-	state.updateRuntimeResolvedTs(resolvedTs, time.Now())
-
-	ts := uint64(0)
-	shouldAdvance := false
-	// advanceInterval defaults to 100ms; setting it to 0 means resolving the timestamp as soon as possible.
-	// Note: If a single span contains an extremely large number of regions (e.g., 500k), advanceInterval = 0 may cause performance issues.
-	if span.advanceInterval == 0 {
-		span.rangeLock.UpdateLockedRangeStateHeap(state.region.lockedRangeState)
-		ts = span.rangeLock.GetHeapMinTs()
-		shouldAdvance = true
-	} else {
-		now := time.Now().UnixMilli()
-		lastAdvance := span.lastAdvanceTime.Load()
-		if now-lastAdvance >= span.advanceInterval && span.lastAdvanceTime.CompareAndSwap(lastAdvance, now) {
-			ts = span.rangeLock.ResolvedTs()
-			shouldAdvance = true
-		}
-	}
-
-	if shouldAdvance {
-		if ts > 0 && span.initialized.CompareAndSwap(false, true) {
-			log.Info("subscription client is initialized",
-				zap.Uint64("subscriptionID", uint64(span.subID)),
-				zap.Uint64("regionID", regionID),
-				zap.Uint64("resolvedTs", ts))
-		}
-		lastResolvedTs := span.resolvedTs.Load()
-		nextResolvedPhyTs := oracle.ExtractPhysical(ts)
-		// Generally, we don't want to send duplicate resolved ts,
-		// so we check whether `ts` is larger than `lastResolvedTs` before send it.
-		// but when `ts` == `lastResolvedTs` == `span.startTs`,
-		// the span may just be initialized and have not receive any resolved ts before,
-		// so we also send ts in this case for quick notification to downstream.
-		if ts > lastResolvedTs || (ts == lastResolvedTs && lastResolvedTs == span.startTs) {
-			resolvedPhyTs := oracle.ExtractPhysical(lastResolvedTs)
-			decreaseLag := float64(nextResolvedPhyTs-resolvedPhyTs) / 1e3
-			const largeResolvedTsAdvanceStepInSecs = 30
-			if decreaseLag > largeResolvedTsAdvanceStepInSecs {
-				log.Warn("resolved ts advance step is too large",
-					zap.Uint64("subID", uint64(span.subID)),
-					zap.Int64("tableID", span.span.TableID),
-					zap.Uint64("regionID", regionID),
-					zap.Uint64("resolvedTs", ts),
-					zap.Uint64("lastResolvedTs", lastResolvedTs),
-					zap.Float64("decreaseLag(s)", decreaseLag))
-			}
-			span.resolvedTs.Store(ts)
-			span.resolvedTsUpdated.Store(time.Now().Unix())
-			return ts
-		}
-	}
-
-	return 0
 }
