+++ logservice/logpuller/subscription_client.go
@@ -207,95 +207,6 @@ type subscriptionClient struct {
 	credential *security.Credential
 }
 
-func (s *subscriptionClient) ensureHelpers() {
-	if s.requestedStores == nil {
-		s.requestedStores = newRequestedStoreSet(s)
-	}
-	if s.regionScheduler == nil {
-		s.regionScheduler = newRegionRequestScheduler(s)
-	}
-	if s.subscribedSpans == nil {
-		s.subscribedSpans = newSubscribedSpanSet(s)
-	}
-	if s.staleLockResolver == nil {
-		s.staleLockResolver = newStaleLockResolver(s)
-	}
-	if s.regionRuntimeRegistry == nil {
-		s.regionRuntimeRegistry = newRegionRuntimeRegistry()
-	}
-	if s.failureStats == nil {
-		s.failureStats = newFailureStats()
-	}
-}
-
-func (s *subscriptionClient) markRegionDiscovered(region *regionInfo, now time.Time) {
-	if region.verID.GetID() == 0 {
-		return
-	}
-	if !region.runtimeKey.isValid() {
-		region.runtimeKey = s.regionRuntimeRegistry.allocKey(region.subscribedSpan.subID, region.verID.GetID())
-		s.regionRuntimeRegistry.markDiscovered(region.runtimeKey, *region, now)
-	}
-}
-
-func (s *subscriptionClient) updateRegionRuntimeInfo(region regionInfo) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.updateRegionInfo(region.runtimeKey, region)
-}
-
-func (s *subscriptionClient) markRegionRangeLockWait(region regionInfo, now time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.markRangeLockWait(region.runtimeKey, now)
-}
-
-func (s *subscriptionClient) markRegionRetryPending(region regionInfo, err error, now time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.markRetryPending(region.runtimeKey, err, now)
-}
-
-func (s *subscriptionClient) markRegionRPCReady(region regionInfo, now time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.markRPCReady(region.runtimeKey, now)
-}
-
-func (s *subscriptionClient) markRegionQueued(region regionInfo, acquiredTime, queuedTime time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.markQueued(region.runtimeKey, acquiredTime, queuedTime, region.resolvedTs())
-}
-
-func (s *subscriptionClient) recordRegionRuntimeError(region regionInfo, err error, now time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.recordError(region.runtimeKey, err, now)
-}
-
-func (s *subscriptionClient) regionRuntimePhaseCounts() map[regionPhase]int {
-	return s.regionRuntimeRegistry.phaseCounts()
-}
-
-func (s *subscriptionClient) removeSubscriptionRuntime(subID SubscriptionID) {
-	s.regionRuntimeRegistry.removeBySubscription(subID)
-}
-
-func (s *subscriptionClient) removeRegionRuntime(region regionInfo, now time.Time) {
-	if !region.runtimeKey.isValid() {
-		return
-	}
-	s.regionRuntimeRegistry.markRemoved(region.runtimeKey, now)
-	s.regionRuntimeRegistry.remove(region.runtimeKey)
-}
-
 // NewSubscriptionClient creates a client.
 func NewSubscriptionClient(
 	config *SubscriptionClientConfig,
@@ -346,64 +257,6 @@ func (s *subscriptionClient) AllocSubscriptionID() SubscriptionID {
 	return SubscriptionID(subscriptionIDGen.Add(1))
 }
 
-func (s *subscriptionClient) updateMetrics(ctx context.Context) error {
-	ticker := time.NewTicker(10 * time.Second)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			resolvedTsLag := s.subscribedSpans.getResolvedTsLag()
-			if resolvedTsLag > 0 {
-				metrics.LogPullerResolvedTsLag.Set(resolvedTsLag)
-			}
-			dsMetrics := s.ds.GetMetrics()
-			metricSubscriptionClientDSChannelSize.Set(float64(dsMetrics.EventChanSize))
-			metricSubscriptionClientDSPendingQueueLen.Set(float64(dsMetrics.PendingQueueLen))
-			if len(dsMetrics.MemoryControl.AreaMemoryMetrics) > 1 {
-				log.Panic("subscription client should have only one area")
-			}
-			if len(dsMetrics.MemoryControl.AreaMemoryMetrics) > 0 {
-				areaMetric := dsMetrics.MemoryControl.AreaMemoryMetrics[0]
-				metrics.DynamicStreamMemoryUsage.WithLabelValues(
-					"log-puller",
-					"max",
-					"default",
-					"default",
-				).Set(float64(areaMetric.MaxMemory()))
-				metrics.DynamicStreamMemoryUsage.WithLabelValues(
-					"log-puller",
-					"used",
-					"default",
-					"default",
-				).Set(float64(areaMetric.MemoryUsage()))
-			}
-
-			requestStats := s.requestedStores.requestStats()
-			metrics.SubscriptionClientRequestedRegionCount.WithLabelValues("pending").Set(float64(requestStats.Total))
-			metrics.SubscriptionClientRequestedRegionCount.WithLabelValues("queued").Set(float64(requestStats.Queued))
-			metrics.SubscriptionClientRequestedRegionCount.WithLabelValues("processing").Set(float64(requestStats.Processing))
-			metrics.SubscriptionClientRequestedRegionCount.WithLabelValues("sent").Set(float64(requestStats.Sent))
-
-			counts := s.regionRuntimePhaseCounts()
-			for _, phase := range regionRuntimePhases {
-				metrics.SubscriptionClientRegionRuntimePhaseCount.WithLabelValues(string(phase)).Set(float64(counts[phase]))
-			}
-			stalledSpanReport := s.collectStalledSpanReport(s.pdClock.CurrentTime(), 0, -1)
-			metrics.SubscriptionClientStalledSpanCount.Set(float64(stalledSpanReport.stalledSpanCount))
-			for _, blockerType := range resolvedTsBlockerTypes {
-				metrics.SubscriptionClientStalledSpanCountByBlockerType.WithLabelValues(string(blockerType)).
-					Set(float64(stalledSpanReport.blockerCounts[blockerType]))
-			}
-			metrics.SubscriptionClientStalledSpanMaxResolvedTsLag.Set(stalledSpanReport.maxResolvedTsLag.Seconds())
-			metrics.SubscriptionClientStalledSpanMaxResolvedTsUpdatedAge.Set(stalledSpanReport.maxResolvedTsUpdatedAgo.Seconds())
-
-			metrics.SubscriptionClientSubscribedRegionCount.Set(float64(s.subscribedSpans.requestedRegionCount()))
-		}
-	}
-}
-
 // Subscribe the given table span.
 // NOTE: `span.TableID` must be set correctly.
 // It creates a subscribed span, registers it, and schedules the bootstrap range request.
@@ -449,56 +302,6 @@ func (s *subscriptionClient) Unsubscribe(subID SubscriptionID) {
 	log.Info("unsubscribe span success",
 		zap.Uint64("subscriptionID", uint64(rt.subID)),
 		zap.Bool("exists", rt != nil))
-}
-
-func (s *subscriptionClient) wakeSubscription(subID SubscriptionID) {
-	s.ds.Wake(subID)
-}
-
-func (s *subscriptionClient) pushRegionEventToDS(subID SubscriptionID, event regionEvent) {
-	// fast path
-	if !s.paused.Load() {
-		s.ds.Push(subID, event)
-		return
-	}
-	// slow path: wait until paused is false
-	s.mu.Lock()
-	for s.paused.Load() {
-		select {
-		case <-s.ctx.Done():
-			s.mu.Unlock()
-			return
-		default:
-			s.cond.Wait()
-		}
-	}
-	s.mu.Unlock()
-	s.ds.Push(subID, event)
-}
-
-func (s *subscriptionClient) handleDSFeedBack(ctx context.Context) error {
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case feedback := <-s.ds.Feedback():
-			switch feedback.FeedbackType {
-			case dynstream.PauseArea:
-				s.mu.Lock()
-				s.paused.Store(true)
-				s.mu.Unlock()
-				log.Info("subscription client pause push region event")
-			case dynstream.ResumeArea:
-				s.mu.Lock()
-				s.paused.Store(false)
-				s.cond.Broadcast()
-				s.mu.Unlock()
-				log.Info("subscription client resume push region event")
-			case dynstream.ReleasePath, dynstream.ResumePath:
-				// Ignore it, because it is no need to pause and resume a path in puller.
-			}
-		}
-	}
 }
 
 func (s *subscriptionClient) Run(ctx context.Context) error {
